# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Insportfolio Api Docs",
        "title": "Insportfolio Docs",
        "contact": {},
        "version": "0.0.1"
    },
    "basePath": "/api/",
    "paths": {
        "/portfolios/": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Portfolios"
                ],
                "summary": "Update profile activated portfolio",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Portfolio URL",
                        "name": "url",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Portfolio preview image",
                        "name": "preview",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "array",
                        "description": "Technologies used",
                        "name": "technologies",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.Portfolio"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "The portfolio is added to the list of portfolios of the profile and the others are setted as active: false",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Portfolios"
                ],
                "summary": "Create portfolio and upload file to AWS",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Portfolio URL",
                        "name": "url",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Portfolio preview image",
                        "name": "preview",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "array",
                        "description": "Technologies used",
                        "name": "technologies",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.Portfolio"
                        }
                    }
                }
            }
        },
        "/portfolios/{profile_id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get portfolio by Profile id, the views counter is incremented only if the user is logged and is not the owner",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Portfolios"
                ],
                "summary": "Get Portfolio by Profile id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "profile_id",
                        "name": "profile_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.GetPortfolioResponse"
                        }
                    }
                }
            }
        },
        "/users/profile": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get a string by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Return authenticated user profile",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemas.ProfileSchema"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "schemas.GetPortfolioResponse": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "comments_count": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "preview": {
                    "type": "string"
                },
                "profile_id": {
                    "type": "integer"
                },
                "reactions_count": {
                    "type": "integer"
                },
                "technologies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schemas.Technology"
                    }
                },
                "updated_at": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "views": {
                    "type": "integer"
                }
            }
        },
        "schemas.Portfolio": {
            "type": "object",
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "preview": {
                    "type": "string"
                },
                "profile_id": {
                    "type": "integer"
                },
                "technologies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schemas.Technology"
                    }
                },
                "updated_at": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "views": {
                    "type": "integer"
                }
            }
        },
        "schemas.ProfileSchema": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "open_to_work": {
                    "type": "boolean"
                },
                "profession": {
                    "type": "string"
                },
                "propic": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "schemas.Technology": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "x-auth-key",
            "in": "header"
        }
    }
}